Package Lock (Dependency Tracking)
There's been some issues with this project for a while now in regards to tracking versions of it's dependencies, I think mostly because they are outdated. I don't want to change the versions for things unsafely, because I don't want to change the project's functionality yet (unintentionally), that will be a later step when I decide to refactor things a bit.